+++ CHANGELOG.md
@@ -12,6 +12,8 @@ omnibus-gitlab repository.
 - Update redis to 2.8.20 6589e23ed79c883988e0ebefc356699f5f94228f
 - Exit on package installation if backup failed and wasn't skipped 710253c318a029bf1bb158c6c9fc81f0f695fe34
 - Added sslmode and sslrootcert database configuration option (Anthony Brodard) dbeb00346ccafdda50e52cf601c6b457b5981b74
+- Added option to disable HTTPS on nginx to support proxied SSL for GitLab CI
+- Added custom listen_port for GitLab CI nginx to support reverse proxies
 
 7.11.0